factor out & encapsulate
